Pulling out index queries into a separate file
